2005-09-08  Martin Baulig  <martin@ximian.com>
	* backends/server/i386-arch.c (server_ptrace_get_target_info):
	Make `target_long_size' 4 bytes, not 8.

	* backends/classes/TargetReader.cs
	(TargetReader.ReadLongInteger): Read either 4 or 8 bytes,
	depending on the `TargetLongIntegerSize'.


svn path=/trunk/debugger/; revision=49715

+++ backends/server/i386-arch.c
@@ -463,7 +463,7 @@ server_ptrace_get_target_info (guint32 *target_int_size, guint32 *target_long_si
 			       guint32 *target_address_size, guint32 *is_bigendian)
 {
 	*target_int_size = sizeof (guint32);
-	*target_long_size = sizeof (guint64);
+	*target_long_size = sizeof (guint32);
 	*target_address_size = sizeof (void *);
 	*is_bigendian = 0;
